Eliminate another postgresql cookbook race condition
Package installation causes PG to start/restart and it takes
time. So when we first do that and only then set up
sysctl values like kernel.shmmax, which operations completes
first is undetermined, but is very important for PG's ability
to start.

So set sysctl values as early as possible.