Hardcode Ruby installation order into rvm::multi
What we are doing here is not cool nor how Chef recipes are typically written but
we have reasons to hardcode Rubies we install: we need them to be installed with various
configuration/installation options and also IN A SPECIFIC ORDER.

If we just use a hash in node attributes, we get the former but not the latter.
This matters for Rubinius which currently needs 1.8.7 to be used when configuring/building it.

A better solution would be to use array of pairs in node attributes. We will go back to
doing this later. This is not as trivial as it sounds:

 * We need to pass custom configuration options

 * We need rubies to be installed in a specific order