Add scripts for deploying CMake and mingw32
For testing the libgit2 windows build, it would be nice if the travis-ci
build VMs had mingw32. It seems to take too long to download/install
sometimes with the slow german ubuntu mirror, or we could just install it
as part of our .travis.yml.

Since we're adding mingw32 anyway, might as well CMake which is also
required for building libgit2.