Install a cron task that makes sure eth0 is up and happy
eth0 with the currently used NIC (virtual network card) sometimes goes down according
to dmesg and syslog. The reason for that seems to be that NICs are overwhelmed by a huge
number of outgoing network connections that tools like pip and Maven may open simultaneously.

ifup eth0 is idempotent so there is no downside to trying this.

This is a yet another step towards eliminating stalling SSH connections. Some of the work
has to be done on the net/ssh side, too.