Add -Xcompile.invokedynamic=false to JRUBY_OPTS
We are trying it out to see if it helps with OpenJDK 6 and 7 segfaults that
ocassionaly happen on JRuby :(

It is something very much 32-bit OpenJDK and JRuby specific: Clojure, Java, Groovy and most of Scala
projects on travis never cause segfaults and 64 bit JVMs don't seem to segfault
either.

Per discussion with @rkh.