Make Oracle JDK 7 the default
Even though the next versions of Debian, Ubuntu and Fedora ship with OpenJDK 7 by default,
and Oracle JDK being based on OpenJDK (so, all new work goes into OpenJDK first), our
OpenJDK 7 package is very old (by recent JDK development standards) and ocassionally
segfaults with JRuby and Scala.

This is a 32-bit problem so our options are:

 * Move CI env to 64 bit
 * Move to 12.04 earlier than expected
 * Upgrade OpenJDK 7 to 7u5

Each of these has problems:

 * We will likely have to upgrade some hardware to make sure we have the same amount of
   memory free for I/O buffers, since 64 bit VMs will use more RAM.

 * There are no 12.04 packages for a number of projects (like PyPy) and we cannot
   quickly build and maintain them well on our own.

 * Building OpenJDK 7 is not that hard but takes ~20 minutes on a powerful machine
   and *packaging* it is a whole different story. We probably won't be able to do
   it and maintain it well on our own either.

So we just switch the default: our Oracle JDK 7 is 7u5 (a recent patch release) and
with 7u6, 7u8 and so on it will be pretty easy for us to upgrade it thanks to the
PPA we use which downloads packages from the Oracle Technology Network mirrors.

We are still committed to helping OpenJDK 7 & 8, but for now we cannot do much about
the situation and a non-trivial portion of our users is affected.