Create 2 sets of virtualenvs, with --system-site-packages enabled and without
There is no way to install some packages via pip but it is a good idea to isolate
from the rest of the system for most projects (that is why --s-s-p is disabled by
default in virtualenv env now).

So the best we can offer is to create two sets and later adapt Python builder
to activate the right one based on an extra .travis.yml attribute.