Make sure hostname is consistent between provisioning
It matters a great deal for Erlang services that use Mnesia (like RabbitMQ).
What happens is that during installation RabbitMQ will create a directory
(a Mnesia DB) with the current hostname in it but then that hostname
is changed during provisioning and on the next boot RabbitMQ fails to
start.