Disable SASL for Memcached, references #31
It was a good idea in theory but does not work out
well in practice. Many Memcached clients do not
support SASL or it is not assumed to be enabled
by default. In addition, most developers do not
even realize Memcached has authentication support
at all. In the end, lots of test suites fail for
no obvious reason.

In addition, authentication is not optional, so
we have to either double down or go back to
running Memcached w/o authentication. The latter
is the only realistic option.