Add the latest Erlang release to ci-environment
The latest, and greatest, Erlang release is now R16B02.
It is also supported by Kerl, but your CI environment
doesn't yet reflect it's release.

This commit adds R16B02 to the list of releases
installed with Kerl.

The documentation for the Erlang setup on Travis-CI
should potentially also be updated to reflect the change.