Removed python 2.5 and upgrade virtualenv to 1.10.1
After talking to @dstufft and @alex, and looking at some download stats from PyPi on the use of Python 2.5, we have decided to remove it from our Python VMs.

https://caremad.io/blog/a-look-at-pypi-downloads/
https://gist.github.com/dstufft/6954117

As Python 2.5 does not work with the most recent VirtualEnv, and that we need to do some annoying tricks to get it to install, and that only 0.29% of PyPi installs came from Python 2.5, the hassle of maintaining the install isn't worth the benefit.

On top of this, as Python 2.5 was stopping us from using the latest VirtualEnv, and thus the latest Pip, it was effectively holding us back from updating the VMs.