Use Ubuntu Java keystore for all JDKs
This way we ensure that:
- all java runtime will behave the same, independently of selected JDK
- most common SSL CA will be supported (including popular authorities like
  StartSSL, ...)

Notes:

clojars.org now uses certificate issued by RapidSSL CA, and I
verified that default keystore provided by Webupd8 packages also work
fine to run Leiningen2. But as argumented above, for any other similar
possible use case, it is safer to use the certificates provided by
Ubuntu packaging.

The current implementation will auto-install an openjdk6 JRE as
dependency of the package ca-certificates-java. This is not *ideal*, but
it should have no impact. As a variant, we could provision a static copy
of the ubuntu keystore file, via cookbook_file or remote_file.