Install PyPy before PyPy3, and create alias pypy3
See https://github.com/travis-ci/travis-ci/issues/2583

By changing the order, the 'pypy' binary in the pypy-2.3.1 will
be installed as 'pypy' in the default `$PATH`, instead of
that found in pypy3-2.3.1.

We also create a new alias `pypy3` which will point to the
_hard-coded_ PyPy3 binary, so that `tox` users can still use it.
There is a [ticket for PyPy](https://bitbucket.org/pypy/pypy/issue/1796/pypy-pypy3-in-the-same-path)
that will make it unnecessary to do this linking, but until then,
we will need this ugly hack.