Tweak libstemmer in Sphinx
Fixes https://github.com/travis-ci/travis-ci/issues/2804

Edit libstemmer's Makefile.in so that it can be used while compiling
Sphinx.
This is a bit of a hack, admittedly. It is not quite clear how
we should address this issue long-term when either libstemmer or Sphinx
make changes, but for now, this is good enough to keep things
going.