Update android-wait-for-emulator for Lollipop
Either adb output changed, or there is a component in Android SDK
that now writes "device offline" instead of "device not found".

Whatever the case, we need to check for this additional output
so that we can get out of the loop.

Additionally, we explicitly wait for the output "running", in order
to get out of the loop.

Finally, the components may need a little longer than 15 seconds
to be ready, so we raise the timeout to 60 seconds.