Rename 'sbt' to 'sbt-extras'
The correct name of this cookbook has always been 'sbt-extras'. With
this change we fix a metadata/directory names mismatch created by
695415b7562e88d72ed9434b6213d2da9b325646.

Note: this change previously would have required to update travis-images
runlists as well, but it is no longer the case as these definitions are now
stored in the same repository, under vm_templates directory!

//cc: @BanzaiMan I am pretty confident that this change should not break
      anything, hence this direct push (no pull request review).