Split io.js installation logic to separate recipe
NVM behaves quite differently when it deals with io.js.
The installation directory under ~/.nvm, and the manner in which
it is referred by the 'nvm' command is quite different.

Rather than dealing with these inconsistencies within 'nodejs::multi'
recipe, we split up this logic to a different recipe.

Conceivably, a Chef provider may be able to deal with such differences
more elegantly (at least from the recipes' POV), but for now,
we allow somewhat duplicated logic here.