Remove unneeded ubuntugis-stable PPA.
Whilst both ubuntugis-stable and pgdg now both use libgdal1h, in the
past ubuntugis-stable used libgdal1, which caused many problems
installing things that depend upon GDAL, see e.g.
https://github.com/travis-ci/travis-ci/issues/2401.

According to the postgresql README, ubuntugis-stable was included to get
libgdal >= 1.9. libgdal 1.9 or later is included in pgdg, so there is no
longer a need for this PPA. Also, pgdg includes the postgresql postgis
packages for 9.4 which it seems ubuntugis-stable does not; another
reason not to use it.