Remove `./node_modules/bin` from `$PATH`
Some Node.js modules provide commands with names identical to well-known commands (including `bash` builtin functions such as `which`). This causes confusing issues when other software assumes that `which` behaves a certain way (see https://github.com/travis-ci/travis-ci/issues/5092).

While we have a good reason to add `./node_modules/bin` to `$PATH`, a consensus among Node.js users is that it shouldn't be.

Removing this from `$PATH` may very well break some builds, but we will address that by writing a message to the logs with `travis-build`. (This will be done separately.)