Bump `android-sdk` cookbook to version 0.2.1
This change is required to install SDK Tools 25+, since Google changed
the repository URL and the packaging extension (.zip instead of .tgz).

- Install Android SDK Tools 25.2.3 and Build Tools 25.0.2
- The other pre-installed components are kept unchanged for backwards
  compatibility reason.

The cookbook directory content corresponds to the package released on
supermarket.chef.io (with manual addition of metadata.rb for sake of
readability, compared to metadata.json).

Note that this commit also removed the two `maven-rescue.rb` duplicates
(of `maven_rescue.rb` files) that were accidentally introduced in
0f497eb71291b52a703143c5cd63a217c8766dc9 (when `site-cookbooks` was
renamed as `community-cookbooks`).