Update docker to 17.06.1 and docker-compose to 1.15.0
For changelogs see:
https://docs.docker.com/release-notes/docker-ce/#17061-ce-2017-08-15
https://docs.docker.com/release-notes/docker-compose/#1150-2017-07-26

As of the docker-ce `17.06.1~ce-0~ubuntu` release, the package name
no longer includes the string `trusty`, so is identical across both
Trusty and Xenial. As such, once this PR is merged, the Xenial override
in packer-templates can be removed:
https://github.com/travis-ci/packer-templates/blob/9c6875413a5395c4c4dbfcbc1eeb7569e3e35b1b/cookbooks/travis_ci_stevonnie/attributes/default.rb#L3

Should the Trusty and Xenial versions ever diverge again, it would be
best to add the conditional to travis-cookbooks rather than overriding
in packer-templates.