Always install the latest stable version of Yarn
The Yarn project is frequently releasing updates, which causes many
requests for the pre-installed version to be bumped. With this PR
the latest version of yarn is always installed at time of image
creation. Ark supports omitting the `checksum` property, which we
have to do since it will be constantly changing. The `version`
property is only used for determining the install location name
(which is then symlinked back to the non-suffixed name), so doesn't
need to be a traditional numeric version string.

Fixes #931.